Move Prefab component type from Sofa_Core bindings to src/SofaPython3 & ADD prefab breaking feature
