+++ bindings/Sofa/src/SofaPython3/Sofa/Core/Binding_Prefab.cpp
@@ -39,11 +39,6 @@ along with sofaqtquick. If not, see <http://www.gnu.org/licenses/>.
 #include <sofa/core/objectmodel/DataCallback.h>
 using sofa::core::objectmodel::DataCallback;
 
-#include <sofa/simulation/DeleteVisitor.h>
-#include <sofa/helper/system/FileMonitor.h>
-using sofa::helper::system::FileMonitor;
-using sofa::helper::system::FileEventListener;
-
 PYBIND11_DECLARE_HOLDER_TYPE(Prefab,
                              sofapython3::py_shared_ptr<Prefab>, true)
 
@@ -55,80 +50,6 @@ using sofa::simulation::Simulation;
 
 namespace sofapython3
 {
-using sofa::core::objectmodel::Event;
-
-void Prefab::init()
-{
-    reinit();
-    Inherit1::init(sofa::core::ExecParams::defaultInstance());
-}
-
-void PrefabFileEventListener::fileHasChanged(const std::string &filename)
-{
-    PythonEnvironment::gil acquire ;
-    //std::string file=filepath;
-    //SP_CALL_FILEFUNC(const_cast<char*>("onReimpAFile"),
-    //                 const_cast<char*>("s"),
-    //                 const_cast<char*>(file.data()));
-
-    py::dict local;
-    local["filename"] = filename;
-    py::eval("onReimpAFile(filename)", py::globals(), local);
-
-    m_prefab->clearLoggedMessages();
-    m_prefab->init();
-}
-
-void Prefab::reinit()
-{
-    clearLoggedMessages();
-
-    /// remove everything in the node.
-    execute<sofa::simulation::DeleteVisitor>(sofa::core::ExecParams::defaultInstance());
-    doReInit();
-
-    /// Beurk beurk beurk
-    sofa::simulation::getSimulation()->initNode(this);
-    execute<VisualInitVisitor>(nullptr);
-
-    m_componentstate = sofa::core::objectmodel::ComponentState::Valid;
-}
-
-void Prefab::doReInit()
-{
-}
-
-Prefab::Prefab()
-{
-    m_filelistener.m_prefab = this;
-    m_datacallback.addCallback( std::bind(&Prefab::reinit, this) );
-}
-
-
-Prefab::~Prefab()
-{
-    FileMonitor::removeListener(&m_filelistener);
-}
-
-
-void Prefab::addPrefabParameter(const std::string& name, py::object value, const std::string& help, std::string type)
-{
-    BaseData* data = findData(name);
-    if(data == nullptr)
-    {
-        BaseData* data = BindingBase::addData(py::cast(this), name, value, py::object(), help, "Prefab's properties", type);
-        m_datacallback.addInputs({data});
-        return;
-    }
-    //PythonFactory::fromPython(data, value);
-}
-
-void Prefab::setSourceTracking(const std::string& filename)
-{
-    std::cout << "Activating source tracking to " << filename << std::endl;
-    FileMonitor::addFile(filename, &m_filelistener);
-}
-
 
 class Prefab_Trampoline : public Prefab, public PythonTrampoline
 {
